fix refresh button padding, issue #19
Set refresh button width to a fixed width matching the width of a standard action bar button for the size and density of the device's screen. This also fixes the lack of padding on either side of the refresh icon in the "RefreshProgressActionView".

Also removes unused import "android.text.AlteredCharSequence" from "AnimeServerInterface.java".